+++ winpr/libwinpr/timezone/TimeZones.c
@@ -1,0 +1,1584 @@
+/*
+ * Automatically generated with scripts/TimeZones.csx
+ */
+
+#pragma pack(push, 1)
+struct _TIME_ZONE_RULE_ENTRY
+{
+	UINT64 TicksStart;
+	UINT64 TicksEnd;
+	INT32 DaylightDelta;
+	SYSTEMTIME StandardDate;
+	SYSTEMTIME DaylightDate;
+};
+typedef struct _TIME_ZONE_RULE_ENTRY TIME_ZONE_RULE_ENTRY;
+
+struct _TIME_ZONE_ENTRY
+{
+	const char* Id;
+	INT32 Bias;
+	BOOL SupportsDST;
+	const char* DisplayName;
+	const char* StandardName;
+	const char* DaylightName;
+	const TIME_ZONE_RULE_ENTRY* RuleTable;
+	UINT32 RuleTableCount;
+};
+typedef struct _TIME_ZONE_ENTRY TIME_ZONE_ENTRY;
+#pragma pack(pop)
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_2[] =
+{
+	{ 633031164000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 4, 0, 1, 2, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 633032028000000000ULL, 60, { 0, 11, 0, 1, 2, 0, 0, 0 }, { 0, 3, 0, 2, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_5[] =
+{
+	{ 633031164000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 4, 0, 1, 2, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 633032028000000000ULL, 60, { 0, 11, 0, 1, 2, 0, 0, 0 }, { 0, 3, 0, 2, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_7[] =
+{
+	{ 633978108000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 4, 0, 1, 2, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 633978972000000000ULL, 60, { 0, 11, 0, 1, 2, 0, 0, 0 }, { 0, 3, 0, 2, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_9[] =
+{
+	{ 633031164000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 4, 0, 1, 2, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 633032028000000000ULL, 60, { 0, 11, 0, 1, 2, 0, 0, 0 }, { 0, 3, 0, 2, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_11[] =
+{
+	{ 3155378076000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 4, 0, 1, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_12[] =
+{
+	{ 633031164000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 4, 0, 1, 2, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 633032028000000000ULL, 60, { 0, 11, 0, 1, 2, 0, 0, 0 }, { 0, 3, 0, 2, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_14[] =
+{
+	{ 633031164000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 4, 0, 1, 2, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 633032028000000000ULL, 60, { 0, 11, 0, 1, 2, 0, 0, 0 }, { 0, 3, 0, 2, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_15[] =
+{
+	{ 633346524000000000ULL, 0ULL, 60, { 0, 3, 6, 2, 22, 0, 0, 0 }, { 0, 10, 6, 2, 22, 0, 0, 0 }, },
+	{ 633662748000000000ULL, 633347388000000000ULL, 60, { 0, 3, 6, 5, 22, 0, 0, 0 }, { 0, 10, 6, 2, 22, 0, 0, 0 }, },
+	{ 633978108000000000ULL, 633663612000000000ULL, 60, { 0, 3, 6, 2, 22, 0, 0, 0 }, { 0, 10, 6, 2, 22, 0, 0, 0 }, },
+	{ 634293468000000000ULL, 633978972000000000ULL, 60, { 0, 4, 6, 1, 22, 0, 0, 0 }, { 0, 10, 6, 2, 22, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 5, 6, 1, 22, 0, 0, 0 }, { 0, 8, 6, 3, 22, 0, 0, 0 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 60, { 0, 4, 6, 5, 22, 0, 0, 0 }, { 0, 9, 6, 1, 22, 0, 0, 0 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 60, { 0, 4, 6, 5, 22, 0, 0, 0 }, { 0, 9, 6, 1, 22, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 4, 6, 5, 22, 0, 0, 0 }, { 0, 9, 6, 1, 22, 0, 0, 0 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 635871996000000000ULL, 60, { 0, 5, 6, 2, 22, 0, 0, 0 }, { 0, 8, 6, 2, 22, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_16[] =
+{
+	{ 3155378076000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 4, 0, 1, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_19[] =
+{
+	{ 635555772000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 4, 0, 1, 2, 0, 0, 0 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 60, { 0, 1, 4, 1, 0, 0, 0, 0 }, { 0, 2, 0, 1, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_20[] =
+{
+	{ 633031164000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 4, 0, 1, 2, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 633032028000000000ULL, 60, { 0, 11, 0, 1, 2, 0, 0, 0 }, { 0, 3, 0, 2, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_21[] =
+{
+	{ 633031164000000000ULL, 0ULL, 60, { 0, 10, 6, 5, 23, 59, 59, 999 }, { 0, 4, 6, 1, 23, 59, 59, 999 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 60, { 0, 11, 0, 1, 2, 0, 0, 0 }, { 0, 3, 0, 2, 2, 0, 0, 0 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 60, { 0, 11, 0, 1, 2, 0, 0, 0 }, { 0, 3, 0, 2, 2, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 11, 0, 1, 2, 0, 0, 0 }, { 0, 3, 0, 2, 2, 0, 0, 0 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 60, { 0, 11, 0, 1, 2, 0, 0, 0 }, { 0, 3, 0, 2, 2, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 636188220000000000ULL, 60, { 0, 11, 0, 1, 2, 0, 0, 0 }, { 0, 3, 0, 2, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_22[] =
+{
+	{ 632084220000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 1, 0, 0, 0 }, { 0, 4, 0, 1, 0, 0, 0, 0 }, },
+	{ 632400444000000000ULL, 632085084000000000ULL, 60, { 0, 1, 4, 1, 0, 0, 0, 0 }, { 0, 3, 0, 5, 0, 0, 0, 0 }, },
+	{ 632715804000000000ULL, 632401308000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 633031164000000000ULL, 632716668000000000ULL, 60, { 0, 10, 0, 5, 1, 0, 0, 0 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 633346524000000000ULL, 633032028000000000ULL, 60, { 0, 10, 0, 5, 1, 0, 0, 0 }, { 0, 3, 0, 2, 0, 0, 0, 0 }, },
+	{ 633662748000000000ULL, 633347388000000000ULL, 60, { 0, 10, 0, 5, 1, 0, 0, 0 }, { 0, 3, 0, 3, 0, 0, 0, 0 }, },
+	{ 633978108000000000ULL, 633663612000000000ULL, 60, { 0, 10, 0, 5, 1, 0, 0, 0 }, { 0, 3, 0, 2, 0, 0, 0, 0 }, },
+	{ 634293468000000000ULL, 633978972000000000ULL, 60, { 0, 10, 0, 5, 1, 0, 0, 0 }, { 0, 3, 0, 2, 0, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 11, 0, 2, 1, 0, 0, 0 }, { 0, 3, 0, 3, 0, 0, 0, 0 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 60, { 0, 11, 0, 1, 1, 0, 0, 0 }, { 0, 4, 0, 1, 0, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 634925916000000000ULL, 60, { 0, 11, 0, 1, 1, 0, 0, 0 }, { 0, 3, 0, 2, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_23[] =
+{
+	{ 633031164000000000ULL, 632716668000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 4, 0, 1, 2, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 633032028000000000ULL, 60, { 0, 11, 0, 1, 2, 0, 0, 0 }, { 0, 3, 0, 2, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_24[] =
+{
+	{ 635555772000000000ULL, 0ULL, 60, { 0, 11, 0, 1, 2, 0, 0, 0 }, { 0, 3, 0, 2, 2, 0, 0, 0 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 60, { 0, 1, 4, 1, 0, 0, 0, 0 }, { 0, 3, 0, 2, 2, 0, 0, 0 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 636502716000000000ULL, 636188220000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 636818076000000000ULL, 636503580000000000ULL, 60, { 0, 11, 0, 1, 2, 0, 0, 0 }, { 0, 1, 1, 1, 0, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 636818940000000000ULL, 60, { 0, 11, 0, 1, 2, 0, 0, 0 }, { 0, 3, 0, 2, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_25[] =
+{
+	{ 633662748000000000ULL, 0ULL, 60, { 0, 3, 6, 2, 23, 59, 59, 999 }, { 0, 10, 6, 3, 23, 59, 59, 999 }, },
+	{ 633978108000000000ULL, 633663612000000000ULL, 60, { 0, 3, 6, 1, 23, 59, 59, 999 }, { 0, 10, 6, 3, 23, 59, 59, 999 }, },
+	{ 634293468000000000ULL, 633978972000000000ULL, 60, { 0, 4, 6, 2, 23, 59, 59, 999 }, { 0, 10, 6, 1, 23, 59, 59, 999 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 4, 6, 2, 23, 59, 59, 999 }, { 0, 10, 6, 1, 23, 59, 59, 999 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 60, { 0, 4, 6, 1, 23, 59, 59, 999 }, { 0, 10, 6, 1, 23, 59, 59, 999 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 60, { 0, 3, 6, 4, 23, 59, 59, 999 }, { 0, 10, 6, 1, 23, 59, 59, 999 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 3, 6, 4, 23, 59, 59, 999 }, { 0, 10, 6, 1, 23, 59, 59, 999 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 60, { 0, 3, 6, 3, 23, 59, 59, 999 }, { 0, 10, 6, 1, 23, 59, 59, 999 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, 60, { 0, 3, 6, 5, 23, 59, 59, 999 }, { 0, 10, 6, 1, 23, 59, 59, 999 }, },
+	{ 636502716000000000ULL, 636188220000000000ULL, 60, { 0, 3, 6, 5, 23, 59, 59, 999 }, { 0, 9, 6, 5, 23, 59, 59, 999 }, },
+	{ 636818076000000000ULL, 636503580000000000ULL, 60, { 0, 3, 6, 4, 23, 59, 59, 999 }, { 0, 10, 6, 1, 23, 59, 59, 999 }, },
+	{ 637133436000000000ULL, 636818940000000000ULL, 60, { 0, 3, 6, 4, 23, 59, 59, 999 }, { 0, 10, 6, 1, 23, 59, 59, 999 }, },
+	{ 637449660000000000ULL, 637134300000000000ULL, 60, { 0, 3, 6, 3, 23, 59, 59, 999 }, { 0, 10, 6, 1, 23, 59, 59, 999 }, },
+	{ 3155378076000000000ULL, 637450524000000000ULL, 60, { 0, 3, 6, 5, 23, 59, 59, 999 }, { 0, 10, 6, 1, 23, 59, 59, 999 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_26[] =
+{
+	{ 633031164000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 4, 0, 1, 2, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 633032028000000000ULL, 60, { 0, 11, 0, 1, 2, 0, 0, 0 }, { 0, 3, 0, 2, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_27[] =
+{
+	{ 633346524000000000ULL, 633032028000000000ULL, 30, { 0, 12, 0, 2, 3, 0, 0, 0 }, { 0, 1, 1, 1, 0, 0, 0, 0 }, },
+	{ 633662748000000000ULL, 633347388000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 633978108000000000ULL, 633663612000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 634293468000000000ULL, 633978972000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, -30, { 0, 5, 0, 1, 2, 30, 0, 0 }, { 0, 1, 5, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_28[] =
+{
+	{ 632400444000000000ULL, 0ULL, 60, { 0, 1, 4, 1, 0, 0, 0, 0 }, { 0, 11, 2, 1, 0, 0, 0, 0 }, },
+	{ 632715804000000000ULL, 632401308000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 0, 3, 0, 0, 0, 0 }, },
+	{ 633031164000000000ULL, 632716668000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 0, 1, 0, 0, 0, 0 }, },
+	{ 633346524000000000ULL, 633032028000000000ULL, 60, { 0, 2, 6, 5, 23, 59, 59, 999 }, { 0, 10, 0, 2, 0, 0, 0, 0 }, },
+	{ 633662748000000000ULL, 633347388000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 0, 3, 0, 0, 0, 0 }, },
+	{ 633978108000000000ULL, 633663612000000000ULL, 60, { 0, 2, 6, 2, 23, 59, 59, 999 }, { 0, 10, 6, 3, 23, 59, 59, 999 }, },
+	{ 634293468000000000ULL, 633978972000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 6, 3, 23, 59, 59, 999 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 6, 3, 23, 59, 59, 999 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 60, { 0, 2, 6, 4, 23, 59, 59, 999 }, { 0, 10, 6, 3, 23, 59, 59, 999 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 6, 3, 23, 59, 59, 999 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 6, 3, 23, 59, 59, 999 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 6, 3, 23, 59, 59, 999 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 6, 3, 23, 59, 59, 999 }, },
+	{ 636502716000000000ULL, 636188220000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 6, 2, 23, 59, 59, 999 }, },
+	{ 636818076000000000ULL, 636503580000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 637133436000000000ULL, 636818940000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 637449660000000000ULL, 637134300000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 6, 5, 23, 59, 59, 999 }, },
+	{ 637765020000000000ULL, 637450524000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 638080380000000000ULL, 637765884000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 638395740000000000ULL, 638081244000000000ULL, 60, { 0, 2, 6, 5, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 638711964000000000ULL, 638396604000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 639027324000000000ULL, 638712828000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 639342684000000000ULL, 639028188000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 6, 5, 23, 59, 59, 999 }, },
+	{ 639658044000000000ULL, 639343548000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 639974268000000000ULL, 639658908000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 640289628000000000ULL, 639975132000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 640604988000000000ULL, 640290492000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 640920348000000000ULL, 640605852000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 641236572000000000ULL, 640921212000000000ULL, 60, { 0, 2, 6, 2, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 641551932000000000ULL, 641237436000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 641867292000000000ULL, 641552796000000000ULL, 60, { 0, 2, 6, 5, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 642182652000000000ULL, 641868156000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 642498876000000000ULL, 642183516000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 642814236000000000ULL, 642499740000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 6, 5, 23, 59, 59, 999 }, },
+	{ 643129596000000000ULL, 642815100000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 643444956000000000ULL, 643130460000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 3155378076000000000ULL, 643445820000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_30[] =
+{
+	{ 633346524000000000ULL, 0ULL, 60, { 0, 3, 6, 2, 23, 59, 59, 999 }, { 0, 10, 6, 2, 23, 59, 59, 999 }, },
+	{ 633662748000000000ULL, 633347388000000000ULL, 60, { 0, 3, 6, 5, 23, 59, 59, 999 }, { 0, 10, 6, 2, 23, 59, 59, 999 }, },
+	{ 633978108000000000ULL, 633663612000000000ULL, 60, { 0, 3, 6, 2, 23, 59, 59, 999 }, { 0, 10, 6, 2, 23, 59, 59, 999 }, },
+	{ 634293468000000000ULL, 633978972000000000ULL, 60, { 0, 4, 6, 1, 23, 59, 59, 999 }, { 0, 10, 6, 2, 23, 59, 59, 999 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 5, 6, 1, 23, 59, 59, 999 }, { 0, 8, 6, 3, 23, 59, 59, 999 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 60, { 0, 4, 6, 5, 23, 59, 59, 999 }, { 0, 9, 6, 1, 23, 59, 59, 999 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 60, { 0, 4, 6, 5, 23, 59, 59, 999 }, { 0, 9, 6, 1, 23, 59, 59, 999 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 4, 6, 5, 23, 59, 59, 999 }, { 0, 9, 6, 1, 23, 59, 59, 999 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 635871996000000000ULL, 60, { 0, 5, 6, 2, 23, 59, 59, 999 }, { 0, 8, 6, 2, 23, 59, 59, 999 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_31[] =
+{
+	{ 633031164000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 0, 1, 0, 0 }, { 0, 4, 0, 1, 0, 1, 0, 0 }, },
+	{ 633346524000000000ULL, 633032028000000000ULL, 60, { 0, 11, 0, 1, 0, 1, 0, 0 }, { 0, 3, 0, 2, 0, 1, 0, 0 }, },
+	{ 633662748000000000ULL, 633347388000000000ULL, 60, { 0, 11, 0, 1, 0, 1, 0, 0 }, { 0, 3, 0, 2, 0, 1, 0, 0 }, },
+	{ 633978108000000000ULL, 633663612000000000ULL, 60, { 0, 11, 0, 1, 0, 1, 0, 0 }, { 0, 3, 0, 2, 0, 1, 0, 0 }, },
+	{ 634293468000000000ULL, 633978972000000000ULL, 60, { 0, 11, 0, 1, 0, 1, 0, 0 }, { 0, 3, 0, 2, 0, 1, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 11, 0, 1, 2, 0, 0, 0 }, { 0, 3, 0, 2, 0, 1, 0, 0 }, },
+	{ 3155378076000000000ULL, 634609692000000000ULL, 60, { 0, 11, 0, 1, 2, 0, 0, 0 }, { 0, 3, 0, 2, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_32[] =
+{
+	{ 634925052000000000ULL, 634609692000000000ULL, 60, { 0, 1, 0, 1, 0, 0, 0, 0 }, { 0, 10, 0, 3, 0, 0, 0, 0 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 1, 2, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_33[] =
+{
+	{ 632400444000000000ULL, 0ULL, 60, { 0, 2, 6, 2, 23, 59, 59, 999 }, { 0, 11, 2, 1, 0, 0, 0, 0 }, },
+	{ 632715804000000000ULL, 632401308000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 0, 3, 0, 0, 0, 0 }, },
+	{ 633031164000000000ULL, 632716668000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 0, 1, 0, 0, 0, 0 }, },
+	{ 633346524000000000ULL, 633032028000000000ULL, 60, { 0, 2, 6, 5, 23, 59, 59, 999 }, { 0, 10, 0, 2, 0, 0, 0, 0 }, },
+	{ 633662748000000000ULL, 633347388000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 0, 3, 0, 0, 0, 0 }, },
+	{ 633978108000000000ULL, 633663612000000000ULL, 60, { 0, 2, 6, 2, 23, 59, 59, 999 }, { 0, 10, 6, 3, 23, 59, 59, 999 }, },
+	{ 634293468000000000ULL, 633978972000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 6, 3, 23, 59, 59, 999 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 6, 3, 23, 59, 59, 999 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 60, { 0, 2, 6, 4, 23, 59, 59, 999 }, { 0, 10, 6, 3, 23, 59, 59, 999 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 6, 3, 23, 59, 59, 999 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 6, 3, 23, 59, 59, 999 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 6, 3, 23, 59, 59, 999 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 6, 3, 23, 59, 59, 999 }, },
+	{ 636502716000000000ULL, 636188220000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 6, 2, 23, 59, 59, 999 }, },
+	{ 636818076000000000ULL, 636503580000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 637133436000000000ULL, 636818940000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 637449660000000000ULL, 637134300000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 6, 5, 23, 59, 59, 999 }, },
+	{ 637765020000000000ULL, 637450524000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 638080380000000000ULL, 637765884000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 638395740000000000ULL, 638081244000000000ULL, 60, { 0, 2, 6, 5, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 638711964000000000ULL, 638396604000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 639027324000000000ULL, 638712828000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 639342684000000000ULL, 639028188000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 6, 5, 23, 59, 59, 999 }, },
+	{ 639658044000000000ULL, 639343548000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 639974268000000000ULL, 639658908000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 640289628000000000ULL, 639975132000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 640604988000000000ULL, 640290492000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 640920348000000000ULL, 640605852000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 641236572000000000ULL, 640921212000000000ULL, 60, { 0, 2, 6, 2, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 641551932000000000ULL, 641237436000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 641867292000000000ULL, 641552796000000000ULL, 60, { 0, 2, 6, 5, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 642182652000000000ULL, 641868156000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 642498876000000000ULL, 642183516000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 642814236000000000ULL, 642499740000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 10, 6, 5, 23, 59, 59, 999 }, },
+	{ 643129596000000000ULL, 642815100000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 643444956000000000ULL, 643130460000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, },
+	{ 3155378076000000000ULL, 643445820000000000ULL, 60, { 0, 2, 6, 3, 23, 59, 59, 999 }, { 0, 11, 6, 1, 23, 59, 59, 999 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_35[] =
+{
+	{ 633346524000000000ULL, 633032028000000000ULL, 60, { 0, 1, 1, 1, 0, 0, 0, 0 }, { 0, 12, 0, 5, 0, 0, 0, 0 }, },
+	{ 633662748000000000ULL, 633347388000000000ULL, 60, { 0, 3, 0, 3, 0, 0, 0, 0 }, { 0, 10, 6, 3, 23, 59, 59, 999 }, },
+	{ 633978108000000000ULL, 633663612000000000ULL, 60, { 0, 3, 6, 2, 23, 59, 59, 999 }, { 0, 1, 4, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_36[] =
+{
+	{ 632400444000000000ULL, 0ULL, 60, { 0, 10, 6, 5, 23, 0, 0, 0 }, { 0, 3, 6, 5, 22, 0, 0, 0 }, },
+	{ 632715804000000000ULL, 632401308000000000ULL, 60, { 0, 10, 6, 5, 23, 0, 0, 0 }, { 0, 3, 6, 5, 22, 0, 0, 0 }, },
+	{ 633031164000000000ULL, 632716668000000000ULL, 60, { 0, 10, 6, 5, 23, 0, 0, 0 }, { 0, 3, 6, 5, 22, 0, 0, 0 }, },
+	{ 633346524000000000ULL, 633032028000000000ULL, 60, { 0, 10, 6, 5, 23, 0, 0, 0 }, { 0, 3, 6, 4, 22, 0, 0, 0 }, },
+	{ 633662748000000000ULL, 633347388000000000ULL, 60, { 0, 10, 6, 5, 23, 0, 0, 0 }, { 0, 3, 6, 5, 22, 0, 0, 0 }, },
+	{ 633978108000000000ULL, 633663612000000000ULL, 60, { 0, 10, 6, 4, 23, 0, 0, 0 }, { 0, 3, 6, 5, 22, 0, 0, 0 }, },
+	{ 634293468000000000ULL, 633978972000000000ULL, 60, { 0, 10, 6, 5, 23, 0, 0, 0 }, { 0, 3, 6, 5, 22, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 10, 6, 5, 23, 0, 0, 0 }, { 0, 3, 6, 5, 22, 0, 0, 0 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 60, { 0, 10, 6, 5, 23, 0, 0, 0 }, { 0, 3, 6, 4, 22, 0, 0, 0 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 60, { 0, 10, 6, 5, 23, 0, 0, 0 }, { 0, 3, 6, 5, 22, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 10, 6, 5, 23, 0, 0, 0 }, { 0, 3, 6, 5, 22, 0, 0, 0 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 60, { 0, 10, 6, 4, 23, 0, 0, 0 }, { 0, 3, 6, 5, 22, 0, 0, 0 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, 60, { 0, 10, 6, 5, 23, 0, 0, 0 }, { 0, 3, 6, 5, 22, 0, 0, 0 }, },
+	{ 636502716000000000ULL, 636188220000000000ULL, 60, { 0, 10, 6, 5, 23, 0, 0, 0 }, { 0, 3, 6, 5, 22, 0, 0, 0 }, },
+	{ 636818076000000000ULL, 636503580000000000ULL, 60, { 0, 10, 6, 5, 23, 0, 0, 0 }, { 0, 3, 6, 4, 22, 0, 0, 0 }, },
+	{ 637133436000000000ULL, 636818940000000000ULL, 60, { 0, 10, 6, 5, 23, 0, 0, 0 }, { 0, 3, 6, 5, 22, 0, 0, 0 }, },
+	{ 637449660000000000ULL, 637134300000000000ULL, 60, { 0, 10, 6, 4, 23, 0, 0, 0 }, { 0, 3, 6, 5, 22, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 637450524000000000ULL, 60, { 0, 10, 6, 5, 23, 0, 0, 0 }, { 0, 3, 6, 5, 22, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_37[] =
+{
+	{ 635555772000000000ULL, 0ULL, 60, { 0, 3, 0, 2, 2, 0, 0, 0 }, { 0, 10, 0, 1, 2, 0, 0, 0 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 60, { 0, 3, 0, 2, 2, 0, 0, 0 }, { 0, 1, 4, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_38[] =
+{
+	{ 633346524000000000ULL, 0ULL, 60, { 0, 3, 6, 2, 23, 59, 59, 999 }, { 0, 10, 6, 2, 23, 59, 59, 999 }, },
+	{ 633662748000000000ULL, 633347388000000000ULL, 60, { 0, 3, 6, 5, 23, 59, 59, 999 }, { 0, 10, 6, 2, 23, 59, 59, 999 }, },
+	{ 633978108000000000ULL, 633663612000000000ULL, 60, { 0, 3, 6, 2, 23, 59, 59, 999 }, { 0, 10, 6, 2, 23, 59, 59, 999 }, },
+	{ 634293468000000000ULL, 633978972000000000ULL, 60, { 0, 4, 6, 1, 23, 59, 59, 999 }, { 0, 10, 6, 2, 23, 59, 59, 999 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 5, 6, 1, 23, 59, 59, 999 }, { 0, 8, 6, 3, 23, 59, 59, 999 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 60, { 0, 4, 6, 5, 23, 59, 59, 999 }, { 0, 9, 6, 1, 23, 59, 59, 999 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 60, { 0, 4, 6, 5, 23, 59, 59, 999 }, { 0, 9, 6, 1, 23, 59, 59, 999 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 4, 6, 5, 23, 59, 59, 999 }, { 0, 9, 6, 1, 23, 59, 59, 999 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, 60, { 0, 5, 6, 2, 23, 59, 59, 999 }, { 0, 8, 6, 2, 23, 59, 59, 999 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_39[] =
+{
+	{ 633031164000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 4, 0, 1, 2, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 633032028000000000ULL, 60, { 0, 11, 0, 1, 2, 0, 0, 0 }, { 0, 3, 0, 2, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_40[] =
+{
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 1, 6, 1, 0, 0, 0, 0 }, { 0, 10, 6, 3, 23, 59, 59, 999 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 60, { 0, 2, 6, 4, 23, 59, 59, 999 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_42[] =
+{
+	{ 3155378076000000000ULL, 0ULL, 60, { 0, 9, 0, 5, 2, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_43[] =
+{
+	{ 3155378076000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 1, 0, 0, 0 }, { 0, 3, 0, 5, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_46[] =
+{
+	{ 633662748000000000ULL, 633347388000000000ULL, 60, { 0, 8, 0, 5, 23, 59, 59, 999 }, { 0, 5, 6, 5, 23, 59, 59, 999 }, },
+	{ 633978108000000000ULL, 633663612000000000ULL, 60, { 0, 8, 4, 3, 23, 59, 59, 999 }, { 0, 5, 0, 5, 23, 59, 59, 999 }, },
+	{ 634293468000000000ULL, 633978972000000000ULL, 60, { 0, 8, 6, 1, 23, 59, 59, 999 }, { 0, 5, 6, 1, 23, 59, 59, 999 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 7, 6, 5, 23, 59, 59, 999 }, { 0, 4, 6, 1, 23, 59, 59, 999 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 60, { 0, 9, 0, 5, 3, 0, 0, 0 }, { 0, 4, 0, 5, 2, 0, 0, 0 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 4, 0, 5, 2, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 636502716000000000ULL, 636188220000000000ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 636818076000000000ULL, 636503580000000000ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 636818940000000000ULL, 60, { 0, 10, 0, 4, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_47[] =
+{
+	{ 3155378076000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 3, 0, 5, 1, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_49[] =
+{
+	{ 3155378076000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_50[] =
+{
+	{ 3155378076000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_51[] =
+{
+	{ 3155378076000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_52[] =
+{
+	{ 636502716000000000ULL, 0ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 636818076000000000ULL, 636503580000000000ULL, -60, { 0, 1, 1, 1, 1, 0, 0, 0 }, { 0, 1, 1, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_53[] =
+{
+	{ 3155378076000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_55[] =
+{
+	{ 634608828000000000ULL, 0ULL, 60, { 0, 10, 5, 5, 1, 0, 0, 0 }, { 0, 3, 4, 5, 23, 59, 59, 999 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 60, { 0, 1, 0, 1, 0, 0, 0, 0 }, { 0, 3, 4, 5, 23, 59, 59, 999 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 60, { 0, 12, 5, 3, 0, 0, 0, 0 }, { 0, 1, 2, 1, 0, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 635241276000000000ULL, 60, { 0, 10, 5, 5, 1, 0, 0, 0 }, { 0, 3, 4, 5, 23, 59, 59, 999 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_56[] =
+{
+	{ 3155378076000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 4, 0, 0, 0 }, { 0, 3, 0, 5, 3, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_57[] =
+{
+	{ 633978108000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 0, 0, 0, 0 }, { 0, 3, 0, 5, 0, 0, 0, 0 }, },
+	{ 634293468000000000ULL, 633978972000000000ULL, 60, { 0, 10, 6, 5, 23, 59, 59, 999 }, { 0, 3, 6, 5, 23, 59, 59, 999 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 10, 6, 5, 23, 59, 59, 999 }, { 0, 3, 6, 5, 23, 59, 59, 999 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 60, { 0, 10, 6, 5, 23, 59, 59, 999 }, { 0, 3, 6, 4, 23, 59, 59, 999 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 60, { 0, 10, 6, 5, 23, 59, 59, 999 }, { 0, 3, 6, 5, 23, 59, 59, 999 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 10, 6, 5, 23, 59, 59, 999 }, { 0, 3, 6, 5, 23, 59, 59, 999 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 60, { 0, 10, 6, 4, 23, 59, 59, 999 }, { 0, 3, 6, 5, 23, 59, 59, 999 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, 60, { 0, 10, 6, 5, 23, 59, 59, 999 }, { 0, 3, 6, 5, 23, 59, 59, 999 }, },
+	{ 636502716000000000ULL, 636188220000000000ULL, 60, { 0, 10, 6, 5, 23, 59, 59, 999 }, { 0, 3, 6, 5, 23, 59, 59, 999 }, },
+	{ 636818076000000000ULL, 636503580000000000ULL, 60, { 0, 10, 6, 5, 23, 59, 59, 999 }, { 0, 3, 6, 4, 23, 59, 59, 999 }, },
+	{ 637133436000000000ULL, 636818940000000000ULL, 60, { 0, 10, 6, 5, 23, 59, 59, 999 }, { 0, 3, 6, 5, 23, 59, 59, 999 }, },
+	{ 637449660000000000ULL, 637134300000000000ULL, 60, { 0, 10, 6, 4, 23, 59, 59, 999 }, { 0, 3, 6, 5, 23, 59, 59, 999 }, },
+	{ 3155378076000000000ULL, 637450524000000000ULL, 60, { 0, 10, 6, 5, 23, 59, 59, 999 }, { 0, 3, 6, 5, 23, 59, 59, 999 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_58[] =
+{
+	{ 632715804000000000ULL, 0ULL, 60, { 0, 9, 4, 5, 23, 59, 59, 999 }, { 0, 4, 5, 5, 0, 0, 0, 0 }, },
+	{ 633031164000000000ULL, 632716668000000000ULL, 60, { 0, 9, 4, 3, 23, 59, 59, 999 }, { 0, 4, 5, 5, 0, 0, 0, 0 }, },
+	{ 633346524000000000ULL, 633032028000000000ULL, 60, { 0, 9, 4, 1, 23, 59, 59, 999 }, { 0, 4, 4, 5, 23, 59, 59, 999 }, },
+	{ 633662748000000000ULL, 633347388000000000ULL, 60, { 0, 8, 4, 5, 23, 59, 59, 999 }, { 0, 4, 4, 5, 23, 59, 59, 999 }, },
+	{ 633978108000000000ULL, 633663612000000000ULL, 60, { 0, 8, 4, 3, 23, 59, 59, 999 }, { 0, 4, 4, 4, 23, 59, 59, 999 }, },
+	{ 634293468000000000ULL, 633978972000000000ULL, 60, { 0, 9, 4, 5, 23, 59, 59, 999 }, { 0, 4, 4, 5, 23, 59, 59, 999 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 9, 4, 5, 23, 59, 59, 999 }, { 0, 5, 4, 3, 23, 59, 59, 999 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_59[] =
+{
+	{ 3155378076000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_60[] =
+{
+	{ 632400444000000000ULL, 0ULL, 60, { 0, 9, 4, 5, 23, 59, 59, 999 }, { 0, 4, 4, 1, 0, 0, 0, 0 }, },
+	{ 632715804000000000ULL, 632401308000000000ULL, 60, { 0, 9, 5, 5, 23, 59, 59, 999 }, { 0, 4, 5, 1, 0, 0, 0, 0 }, },
+	{ 633031164000000000ULL, 632716668000000000ULL, 60, { 0, 9, 4, 3, 23, 59, 59, 999 }, { 0, 4, 6, 1, 0, 0, 0, 0 }, },
+	{ 633346524000000000ULL, 633032028000000000ULL, 60, { 0, 11, 4, 1, 23, 59, 59, 999 }, { 0, 3, 5, 5, 0, 0, 0, 0 }, },
+	{ 633662748000000000ULL, 633347388000000000ULL, 60, { 0, 10, 5, 5, 23, 59, 59, 999 }, { 0, 4, 5, 1, 0, 0, 0, 0 }, },
+	{ 633978108000000000ULL, 633663612000000000ULL, 60, { 0, 10, 4, 5, 23, 59, 59, 999 }, { 0, 3, 5, 5, 0, 0, 0, 0 }, },
+	{ 634293468000000000ULL, 633978972000000000ULL, 60, { 0, 10, 4, 5, 23, 59, 59, 999 }, { 0, 4, 5, 1, 0, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 10, 4, 5, 23, 59, 59, 999 }, { 0, 4, 5, 1, 0, 0, 0, 0 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 60, { 0, 10, 4, 5, 23, 59, 59, 999 }, { 0, 3, 5, 5, 0, 0, 0, 0 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 60, { 0, 10, 4, 4, 23, 59, 59, 999 }, { 0, 3, 5, 5, 0, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 10, 4, 5, 23, 59, 59, 999 }, { 0, 3, 5, 5, 0, 0, 0, 0 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 60, { 0, 10, 4, 5, 23, 59, 59, 999 }, { 0, 3, 5, 5, 0, 0, 0, 0 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, 60, { 0, 10, 4, 5, 23, 59, 59, 999 }, { 0, 3, 5, 5, 0, 0, 0, 0 }, },
+	{ 636502716000000000ULL, 636188220000000000ULL, 60, { 0, 10, 4, 5, 23, 59, 59, 999 }, { 0, 3, 5, 5, 0, 0, 0, 0 }, },
+	{ 636818076000000000ULL, 636503580000000000ULL, 60, { 0, 10, 4, 5, 23, 59, 59, 999 }, { 0, 3, 5, 5, 0, 0, 0, 0 }, },
+	{ 637133436000000000ULL, 636818940000000000ULL, 60, { 0, 10, 4, 4, 23, 59, 59, 999 }, { 0, 3, 5, 5, 0, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 637134300000000000ULL, 60, { 0, 10, 4, 5, 23, 59, 59, 999 }, { 0, 3, 5, 5, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_61[] =
+{
+	{ 634925052000000000ULL, 0ULL, 60, { 0, 9, 5, 3, 1, 0, 0, 0 }, { 0, 3, 4, 5, 23, 59, 59, 999 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 60, { 0, 9, 4, 5, 23, 59, 59, 999 }, { 0, 3, 4, 5, 23, 59, 59, 999 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 10, 4, 4, 23, 59, 59, 999 }, { 0, 3, 4, 5, 23, 59, 59, 999 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 60, { 0, 10, 4, 4, 23, 59, 59, 999 }, { 0, 3, 5, 5, 23, 59, 59, 999 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, 60, { 0, 10, 6, 5, 1, 0, 0, 0 }, { 0, 3, 6, 5, 1, 0, 0, 0 }, },
+	{ 636502716000000000ULL, 636188220000000000ULL, 60, { 0, 10, 6, 5, 1, 0, 0, 0 }, { 0, 3, 6, 5, 1, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 636503580000000000ULL, 60, { 0, 10, 6, 5, 1, 0, 0, 0 }, { 0, 3, 6, 4, 1, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_63[] =
+{
+	{ 3155378076000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 4, 0, 0, 0 }, { 0, 3, 0, 5, 3, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_64[] =
+{
+	{ 632400444000000000ULL, 0ULL, 60, { 0, 9, 3, 4, 1, 0, 0, 0 }, { 0, 4, 3, 1, 1, 0, 0, 0 }, },
+	{ 632715804000000000ULL, 632401308000000000ULL, 60, { 0, 10, 0, 2, 2, 0, 0, 0 }, { 0, 4, 5, 1, 2, 0, 0, 0 }, },
+	{ 633031164000000000ULL, 632716668000000000ULL, 60, { 0, 10, 0, 1, 2, 0, 0, 0 }, { 0, 3, 5, 5, 2, 0, 0, 0 }, },
+	{ 633346524000000000ULL, 633032028000000000ULL, 60, { 0, 9, 0, 3, 2, 0, 0, 0 }, { 0, 3, 5, 5, 2, 0, 0, 0 }, },
+	{ 633662748000000000ULL, 633347388000000000ULL, 60, { 0, 10, 0, 1, 2, 0, 0, 0 }, { 0, 3, 5, 5, 2, 0, 0, 0 }, },
+	{ 633978108000000000ULL, 633663612000000000ULL, 60, { 0, 9, 0, 5, 2, 0, 0, 0 }, { 0, 3, 5, 5, 2, 0, 0, 0 }, },
+	{ 634293468000000000ULL, 633978972000000000ULL, 60, { 0, 9, 0, 2, 2, 0, 0, 0 }, { 0, 3, 5, 5, 2, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 10, 0, 1, 2, 0, 0, 0 }, { 0, 4, 5, 1, 2, 0, 0, 0 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 60, { 0, 9, 0, 4, 2, 0, 0, 0 }, { 0, 3, 5, 5, 2, 0, 0, 0 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 3, 5, 5, 2, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 3, 5, 5, 2, 0, 0, 0 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 3, 5, 5, 2, 0, 0, 0 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 3, 5, 5, 2, 0, 0, 0 }, },
+	{ 636502716000000000ULL, 636188220000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 3, 5, 4, 2, 0, 0, 0 }, },
+	{ 636818076000000000ULL, 636503580000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 3, 5, 4, 2, 0, 0, 0 }, },
+	{ 637133436000000000ULL, 636818940000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 3, 5, 5, 2, 0, 0, 0 }, },
+	{ 637449660000000000ULL, 637134300000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 3, 5, 5, 2, 0, 0, 0 }, },
+	{ 637765020000000000ULL, 637450524000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 3, 5, 5, 2, 0, 0, 0 }, },
+	{ 638080380000000000ULL, 637765884000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 3, 5, 5, 2, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 638081244000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 3, 5, 4, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_65[] =
+{
+	{ 634293468000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 1, 6, 1, 0, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 1, 3, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_66[] =
+{
+	{ 636187356000000000ULL, 0ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 636502716000000000ULL, 636188220000000000ULL, 60, { 0, 10, 2, 5, 23, 59, 59, 999 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_67[] =
+{
+	{ 634925052000000000ULL, 634609692000000000ULL, 60, { 0, 11, 6, 2, 2, 0, 0, 0 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 60, { 0, 1, 2, 1, 0, 0, 0, 0 }, { 0, 3, 5, 5, 1, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_68[] =
+{
+	{ 636502716000000000ULL, 0ULL, 60, { 0, 4, 0, 1, 2, 0, 0, 0 }, { 0, 9, 0, 1, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_69[] =
+{
+	{ 632400444000000000ULL, 0ULL, 60, { 0, 10, 5, 1, 4, 0, 0, 0 }, { 0, 4, 4, 1, 3, 0, 0, 0 }, },
+	{ 632715804000000000ULL, 632401308000000000ULL, 60, { 0, 10, 6, 1, 4, 0, 0, 0 }, { 0, 4, 5, 1, 3, 0, 0, 0 }, },
+	{ 633031164000000000ULL, 632716668000000000ULL, 60, { 0, 10, 0, 1, 4, 0, 0, 0 }, { 0, 4, 6, 1, 3, 0, 0, 0 }, },
+	{ 633346524000000000ULL, 633032028000000000ULL, 60, { 0, 10, 1, 1, 4, 0, 0, 0 }, { 0, 4, 0, 1, 3, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_70[] =
+{
+	{ 634293468000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 4, 0, 0, 0 }, { 0, 3, 0, 5, 3, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 10, 0, 5, 4, 0, 0, 0 }, { 0, 3, 1, 5, 3, 0, 0, 0 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 60, { 0, 10, 0, 5, 4, 0, 0, 0 }, { 0, 3, 0, 5, 3, 0, 0, 0 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 60, { 0, 10, 0, 5, 4, 0, 0, 0 }, { 0, 3, 0, 5, 3, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 10, 0, 5, 4, 0, 0, 0 }, { 0, 3, 1, 5, 3, 0, 0, 0 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 60, { 0, 11, 0, 2, 4, 0, 0, 0 }, { 0, 3, 0, 5, 3, 0, 0, 0 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, -60, { 0, 3, 0, 5, 3, 0, 0, 0 }, { 0, 1, 5, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_72[] =
+{
+	{ 634293468000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 1, 6, 1, 0, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_73[] =
+{
+	{ 634293468000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 1, 6, 1, 0, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 1, 3, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_75[] =
+{
+	{ 632400444000000000ULL, 0ULL, 60, { 0, 9, 1, 3, 23, 59, 59, 999 }, { 0, 3, 0, 3, 0, 0, 0, 0 }, },
+	{ 632715804000000000ULL, 632401308000000000ULL, 60, { 0, 9, 3, 3, 23, 59, 59, 999 }, { 0, 3, 2, 4, 0, 0, 0, 0 }, },
+	{ 633662748000000000ULL, 633347388000000000ULL, 60, { 0, 9, 6, 3, 23, 59, 59, 999 }, { 0, 3, 5, 3, 0, 0, 0, 0 }, },
+	{ 633978108000000000ULL, 633663612000000000ULL, 60, { 0, 9, 1, 3, 23, 59, 59, 999 }, { 0, 3, 0, 4, 0, 0, 0, 0 }, },
+	{ 634293468000000000ULL, 633978972000000000ULL, 60, { 0, 9, 2, 3, 23, 59, 59, 999 }, { 0, 3, 1, 4, 0, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 9, 3, 3, 23, 59, 59, 999 }, { 0, 3, 2, 4, 0, 0, 0, 0 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 60, { 0, 9, 4, 3, 23, 59, 59, 999 }, { 0, 3, 3, 3, 0, 0, 0, 0 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 60, { 0, 9, 6, 3, 23, 59, 59, 999 }, { 0, 3, 5, 4, 0, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 9, 0, 3, 23, 59, 59, 999 }, { 0, 3, 6, 4, 0, 0, 0, 0 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 60, { 0, 9, 1, 3, 23, 59, 59, 999 }, { 0, 3, 0, 4, 0, 0, 0, 0 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, 60, { 0, 9, 2, 3, 23, 59, 59, 999 }, { 0, 3, 1, 3, 0, 0, 0, 0 }, },
+	{ 636502716000000000ULL, 636188220000000000ULL, 60, { 0, 9, 4, 3, 23, 59, 59, 999 }, { 0, 3, 3, 4, 0, 0, 0, 0 }, },
+	{ 636818076000000000ULL, 636503580000000000ULL, 60, { 0, 9, 5, 3, 23, 59, 59, 999 }, { 0, 3, 4, 4, 0, 0, 0, 0 }, },
+	{ 637133436000000000ULL, 636818940000000000ULL, 60, { 0, 9, 6, 3, 23, 59, 59, 999 }, { 0, 3, 5, 4, 0, 0, 0, 0 }, },
+	{ 637449660000000000ULL, 637134300000000000ULL, 60, { 0, 9, 0, 3, 23, 59, 59, 999 }, { 0, 3, 6, 3, 0, 0, 0, 0 }, },
+	{ 637765020000000000ULL, 637450524000000000ULL, 60, { 0, 9, 2, 3, 23, 59, 59, 999 }, { 0, 3, 1, 4, 0, 0, 0, 0 }, },
+	{ 638080380000000000ULL, 637765884000000000ULL, 60, { 0, 9, 3, 3, 23, 59, 59, 999 }, { 0, 3, 2, 4, 0, 0, 0, 0 }, },
+	{ 638395740000000000ULL, 638081244000000000ULL, 60, { 0, 9, 4, 3, 23, 59, 59, 999 }, { 0, 3, 3, 4, 0, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 638396604000000000ULL, 60, { 0, 9, 5, 3, 23, 59, 59, 999 }, { 0, 3, 4, 3, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_77[] =
+{
+	{ 634293468000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 1, 6, 1, 0, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 1, 3, 1, 0, 0, 0, 0 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, -60, { 0, 3, 0, 5, 2, 0, 0, 0 }, { 0, 1, 5, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_78[] =
+{
+	{ 635871132000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 5, 0, 0, 0 }, { 0, 3, 0, 5, 4, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_79[] =
+{
+	{ 633978108000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634293468000000000ULL, 633978972000000000ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 1, 5, 1, 0, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 1, 6, 1, 0, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_80[] =
+{
+	{ 633662748000000000ULL, 633347388000000000ULL, 60, { 0, 1, 2, 1, 0, 0, 0, 0 }, { 0, 10, 0, 5, 2, 0, 0, 0 }, },
+	{ 633978108000000000ULL, 633663612000000000ULL, 60, { 0, 3, 0, 5, 2, 0, 0, 0 }, { 0, 1, 4, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_81[] =
+{
+	{ 634293468000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 1, 6, 1, 0, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 1, 3, 1, 0, 0, 0, 0 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, -60, { 0, 12, 0, 1, 2, 0, 0, 0 }, { 0, 1, 5, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_83[] =
+{
+	{ 634608828000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_86[] =
+{
+	{ 634293468000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 1, 6, 1, 0, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 1, 3, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_87[] =
+{
+	{ 633662748000000000ULL, 633347388000000000ULL, 60, { 0, 10, 5, 5, 23, 59, 59, 999 }, { 0, 5, 6, 5, 23, 59, 59, 999 }, },
+	{ 633978108000000000ULL, 633663612000000000ULL, 60, { 0, 10, 6, 5, 23, 59, 59, 999 }, { 0, 4, 2, 2, 23, 59, 59, 999 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_92[] =
+{
+	{ 633978108000000000ULL, 633663612000000000ULL, 60, { 0, 12, 4, 5, 23, 59, 59, 999 }, { 0, 6, 5, 3, 23, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_93[] =
+{
+	{ 634293468000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 1, 6, 1, 0, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 1, 3, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_96[] =
+{
+	{ 634293468000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 1, 6, 1, 0, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 1, 3, 1, 0, 0, 0, 0 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, -60, { 0, 3, 0, 5, 2, 0, 0, 0 }, { 0, 1, 5, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_97[] =
+{
+	{ 633031164000000000ULL, 0ULL, 60, { 0, 9, 6, 5, 2, 0, 0, 0 }, { 0, 3, 6, 5, 2, 0, 0, 0 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 60, { 0, 9, 5, 5, 23, 59, 59, 999 }, { 0, 3, 6, 5, 2, 0, 0, 0 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, 60, { 0, 9, 5, 4, 23, 59, 59, 999 }, { 0, 3, 6, 5, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_98[] =
+{
+	{ 634293468000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 1, 6, 1, 0, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 1, 3, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_99[] =
+{
+	{ 634293468000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 1, 6, 1, 0, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 1, 3, 1, 0, 0, 0, 0 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, -60, { 0, 7, 0, 4, 2, 0, 0, 0 }, { 0, 1, 5, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_100[] =
+{
+	{ 634293468000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 1, 6, 1, 0, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 1, 3, 1, 0, 0, 0, 0 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, -60, { 0, 5, 0, 5, 2, 0, 0, 0 }, { 0, 1, 5, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_102[] =
+{
+	{ 634293468000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 1, 6, 1, 0, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 1, 3, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_104[] =
+{
+	{ 633031164000000000ULL, 632716668000000000ULL, 60, { 0, 1, 0, 1, 0, 0, 0, 0 }, { 0, 12, 0, 1, 2, 0, 0, 0 }, },
+	{ 633346524000000000ULL, 633032028000000000ULL, 60, { 0, 3, 0, 5, 3, 0, 0, 0 }, { 0, 10, 0, 5, 2, 0, 0, 0 }, },
+	{ 633662748000000000ULL, 633347388000000000ULL, 60, { 0, 3, 0, 5, 3, 0, 0, 0 }, { 0, 10, 0, 5, 2, 0, 0, 0 }, },
+	{ 633978108000000000ULL, 633663612000000000ULL, 60, { 0, 3, 0, 5, 3, 0, 0, 0 }, { 0, 1, 4, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_106[] =
+{
+	{ 635871132000000000ULL, 635556636000000000ULL, 60, { 0, 9, 5, 5, 23, 59, 59, 999 }, { 0, 3, 6, 5, 2, 0, 0, 0 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, 60, { 0, 9, 5, 4, 23, 59, 59, 999 }, { 0, 3, 6, 5, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_108[] =
+{
+	{ 634293468000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 1, 6, 1, 0, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 120, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 1, 3, 1, 0, 0, 0, 0 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, -60, { 0, 3, 0, 5, 2, 0, 0, 0 }, { 0, 1, 5, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_110[] =
+{
+	{ 635871132000000000ULL, 635556636000000000ULL, 30, { 0, 8, 5, 2, 23, 59, 59, 999 }, { 0, 1, 4, 1, 0, 0, 0, 0 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 636502716000000000ULL, 636188220000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 636818076000000000ULL, 636503580000000000ULL, -30, { 0, 5, 5, 1, 23, 30, 0, 0 }, { 0, 1, 1, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_112[] =
+{
+	{ 634293468000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 1, 6, 1, 0, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 1, 3, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_113[] =
+{
+	{ 633346524000000000ULL, 0ULL, 60, { 0, 3, 0, 5, 3, 0, 0, 0 }, { 0, 10, 0, 5, 2, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 633347388000000000ULL, 60, { 0, 4, 0, 1, 3, 0, 0, 0 }, { 0, 10, 0, 1, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_116[] =
+{
+	{ 633346524000000000ULL, 0ULL, 60, { 0, 3, 0, 5, 3, 0, 0, 0 }, { 0, 10, 0, 5, 2, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 633347388000000000ULL, 60, { 0, 4, 0, 1, 3, 0, 0, 0 }, { 0, 10, 0, 1, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_118[] =
+{
+	{ 633346524000000000ULL, 0ULL, 60, { 0, 3, 0, 5, 3, 0, 0, 0 }, { 0, 10, 0, 1, 2, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 633347388000000000ULL, 60, { 0, 4, 0, 1, 3, 0, 0, 0 }, { 0, 10, 0, 1, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_119[] =
+{
+	{ 634293468000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 1, 6, 1, 0, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 1, 3, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_120[] =
+{
+	{ 632715804000000000ULL, 0ULL, 30, { 0, 3, 0, 5, 2, 0, 0, 0 }, { 0, 10, 0, 5, 2, 0, 0, 0 }, },
+	{ 633031164000000000ULL, 632716668000000000ULL, 30, { 0, 4, 0, 1, 2, 0, 0, 0 }, { 0, 10, 0, 5, 2, 0, 0, 0 }, },
+	{ 633346524000000000ULL, 633032028000000000ULL, 30, { 0, 3, 0, 5, 2, 0, 0, 0 }, { 0, 10, 0, 5, 2, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 633347388000000000ULL, 30, { 0, 4, 0, 1, 2, 0, 0, 0 }, { 0, 10, 0, 1, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_121[] =
+{
+	{ 635240412000000000ULL, 0ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 1, 3, 1, 0, 0, 0, 0 }, { 0, 12, 0, 5, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_122[] =
+{
+	{ 634293468000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 1, 6, 1, 0, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 1, 3, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_123[] =
+{
+	{ 634293468000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 1, 6, 1, 0, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 120, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 1, 3, 1, 0, 0, 0, 0 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, -60, { 0, 4, 0, 5, 2, 0, 0, 0 }, { 0, 1, 5, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_124[] =
+{
+	{ 635555772000000000ULL, 0ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 30, { 0, 10, 0, 1, 2, 0, 0, 0 }, { 0, 1, 4, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_125[] =
+{
+	{ 634293468000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 1, 6, 1, 0, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 10, 0, 5, 2, 0, 0, 0 }, { 0, 1, 3, 1, 0, 0, 0, 0 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, -60, { 0, 3, 0, 5, 2, 0, 0, 0 }, { 0, 1, 5, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_127[] =
+{
+	{ 633978108000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, },
+	{ 634293468000000000ULL, 633978972000000000ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 1, 5, 1, 0, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 1, 6, 1, 0, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_128[] =
+{
+	{ 633031164000000000ULL, 0ULL, 60, { 0, 3, 0, 3, 3, 0, 0, 0 }, { 0, 10, 0, 1, 2, 0, 0, 0 }, },
+	{ 633346524000000000ULL, 633032028000000000ULL, 60, { 0, 3, 0, 3, 3, 0, 0, 0 }, { 0, 9, 0, 5, 2, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 633347388000000000ULL, 60, { 0, 4, 0, 1, 3, 0, 0, 0 }, { 0, 9, 0, 5, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_130[] =
+{
+	{ 633978108000000000ULL, 633663612000000000ULL, 60, { 0, 1, 4, 1, 0, 0, 0, 0 }, { 0, 11, 0, 5, 2, 0, 0, 0 }, },
+	{ 634293468000000000ULL, 633978972000000000ULL, 60, { 0, 3, 0, 5, 3, 0, 0, 0 }, { 0, 10, 0, 4, 2, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 3, 0, 1, 3, 0, 0, 0 }, { 0, 10, 0, 4, 2, 0, 0, 0 }, },
+	{ 634925052000000000ULL, 634609692000000000ULL, 60, { 0, 1, 0, 4, 3, 0, 0, 0 }, { 0, 10, 0, 3, 2, 0, 0, 0 }, },
+	{ 635240412000000000ULL, 634925916000000000ULL, 60, { 0, 1, 0, 3, 3, 0, 0, 0 }, { 0, 10, 0, 4, 2, 0, 0, 0 }, },
+	{ 635555772000000000ULL, 635241276000000000ULL, 60, { 0, 1, 0, 3, 2, 0, 0, 0 }, { 0, 11, 0, 1, 2, 0, 0, 0 }, },
+	{ 635871132000000000ULL, 635556636000000000ULL, 60, { 0, 1, 0, 3, 3, 0, 0, 0 }, { 0, 11, 0, 1, 2, 0, 0, 0 }, },
+	{ 636187356000000000ULL, 635871996000000000ULL, 60, { 0, 1, 0, 3, 3, 0, 0, 0 }, { 0, 11, 0, 1, 2, 0, 0, 0 }, },
+	{ 636502716000000000ULL, 636188220000000000ULL, 60, { 0, 1, 0, 3, 3, 0, 0, 0 }, { 0, 11, 0, 1, 2, 0, 0, 0 }, },
+	{ 636818076000000000ULL, 636503580000000000ULL, 60, { 0, 1, 0, 2, 3, 0, 0, 0 }, { 0, 11, 0, 1, 2, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 636818940000000000ULL, 60, { 0, 1, 0, 3, 3, 0, 0, 0 }, { 0, 11, 0, 1, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_131[] =
+{
+	{ 3155378076000000000ULL, 0ULL, 60, { 0, 10, 0, 5, 3, 0, 0, 0 }, { 0, 3, 0, 5, 2, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_132[] =
+{
+	{ 633031164000000000ULL, 0ULL, 60, { 0, 3, 0, 3, 3, 45, 0, 0 }, { 0, 10, 0, 1, 2, 45, 0, 0 }, },
+	{ 633346524000000000ULL, 633032028000000000ULL, 60, { 0, 3, 0, 3, 3, 45, 0, 0 }, { 0, 9, 0, 5, 2, 45, 0, 0 }, },
+	{ 3155378076000000000ULL, 633347388000000000ULL, 60, { 0, 4, 0, 1, 3, 45, 0, 0 }, { 0, 9, 0, 5, 2, 45, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_134[] =
+{
+	{ 636187356000000000ULL, 635871996000000000ULL, 60, { 0, 1, 5, 1, 0, 0, 0, 0 }, { 0, 11, 0, 1, 2, 0, 0, 0 }, },
+	{ 636502716000000000ULL, 636188220000000000ULL, 60, { 0, 1, 0, 3, 3, 0, 0, 0 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_RULE_ENTRY TimeZoneRuleTable_135[] =
+{
+	{ 633978108000000000ULL, 0ULL, 0, { 0, 1, 0, 1, 0, 0, 0, 1 }, { 0, 1, 0, 1, 0, 0, 0, 0 }, },
+	{ 634293468000000000ULL, 633978972000000000ULL, 60, { 0, 1, 5, 1, 0, 0, 0, 0 }, { 0, 9, 0, 5, 0, 0, 0, 0 }, },
+	{ 634608828000000000ULL, 634294332000000000ULL, 60, { 0, 4, 6, 1, 4, 0, 0, 0 }, { 0, 9, 6, 4, 3, 0, 0, 0 }, },
+	{ 3155378076000000000ULL, 634609692000000000ULL, 60, { 0, 4, 0, 1, 4, 0, 0, 0 }, { 0, 9, 0, 5, 3, 0, 0, 0 }, }
+};
+
+static const TIME_ZONE_ENTRY TimeZoneTable[] =
+{
+	{
+		"Dateline Standard Time", 720, FALSE, "(UTC-12:00) International Date Line West",
+		"Dateline Standard Time", "Dateline Daylight Time",
+		NULL, 0
+	},
+	{
+		"UTC-11", 660, FALSE, "(UTC-11:00) Coordinated Universal Time-11",
+		"UTC-11", "UTC-11",
+		NULL, 0
+	},
+	{
+		"Aleutian Standard Time", 600, TRUE, "(UTC-10:00) Aleutian Islands",
+		"Aleutian Standard Time", "Aleutian Daylight Time",
+		TimeZoneRuleTable_2, 2
+	},
+	{
+		"Hawaiian Standard Time", 600, FALSE, "(UTC-10:00) Hawaii",
+		"Hawaiian Standard Time", "Hawaiian Daylight Time",
+		NULL, 0
+	},
+	{
+		"Marquesas Standard Time", 570, FALSE, "(UTC-09:30) Marquesas Islands",
+		"Marquesas Standard Time", "Marquesas Daylight Time",
+		NULL, 0
+	},
+	{
+		"Alaskan Standard Time", 540, TRUE, "(UTC-09:00) Alaska",
+		"Alaskan Standard Time", "Alaskan Daylight Time",
+		TimeZoneRuleTable_5, 2
+	},
+	{
+		"UTC-09", 540, FALSE, "(UTC-09:00) Coordinated Universal Time-09",
+		"UTC-09", "UTC-09",
+		NULL, 0
+	},
+	{
+		"Pacific Standard Time (Mexico)", 480, TRUE, "(UTC-08:00) Baja California",
+		"Pacific Standard Time (Mexico)", "Pacific Daylight Time (Mexico)",
+		TimeZoneRuleTable_7, 2
+	},
+	{
+		"UTC-08", 480, FALSE, "(UTC-08:00) Coordinated Universal Time-08",
+		"UTC-08", "UTC-08",
+		NULL, 0
+	},
+	{
+		"Pacific Standard Time", 480, TRUE, "(UTC-08:00) Pacific Time (US & Canada)",
+		"Pacific Standard Time", "Pacific Daylight Time",
+		TimeZoneRuleTable_9, 2
+	},
+	{
+		"US Mountain Standard Time", 420, FALSE, "(UTC-07:00) Arizona",
+		"US Mountain Standard Time", "US Mountain Daylight Time",
+		NULL, 0
+	},
+	{
+		"Mountain Standard Time (Mexico)", 420, TRUE, "(UTC-07:00) Chihuahua, La Paz, Mazatlan",
+		"Mountain Standard Time (Mexico)", "Mountain Daylight Time (Mexico)",
+		TimeZoneRuleTable_11, 1
+	},
+	{
+		"Mountain Standard Time", 420, TRUE, "(UTC-07:00) Mountain Time (US & Canada)",
+		"Mountain Standard Time", "Mountain Daylight Time",
+		TimeZoneRuleTable_12, 2
+	},
+	{
+		"Central America Standard Time", 360, FALSE, "(UTC-06:00) Central America",
+		"Central America Standard Time", "Central America Daylight Time",
+		NULL, 0
+	},
+	{
+		"Central Standard Time", 360, TRUE, "(UTC-06:00) Central Time (US & Canada)",
+		"Central Standard Time", "Central Daylight Time",
+		TimeZoneRuleTable_14, 2
+	},
+	{
+		"Easter Island Standard Time", 360, TRUE, "(UTC-06:00) Easter Island",
+		"Easter Island Standard Time", "Easter Island Daylight Time",
+		TimeZoneRuleTable_15, 10
+	},
+	{
+		"Central Standard Time (Mexico)", 360, TRUE, "(UTC-06:00) Guadalajara, Mexico City, Monterrey",
+		"Central Standard Time (Mexico)", "Central Daylight Time (Mexico)",
+		TimeZoneRuleTable_16, 1
+	},
+	{
+		"Canada Central Standard Time", 360, FALSE, "(UTC-06:00) Saskatchewan",
+		"Canada Central Standard Time", "Canada Central Daylight Time",
+		NULL, 0
+	},
+	{
+		"SA Pacific Standard Time", 300, FALSE, "(UTC-05:00) Bogota, Lima, Quito, Rio Branco",
+		"SA Pacific Standard Time", "SA Pacific Daylight Time",
+		NULL, 0
+	},
+	{
+		"Eastern Standard Time (Mexico)", 300, TRUE, "(UTC-05:00) Chetumal",
+		"Eastern Standard Time (Mexico)", "Eastern Daylight Time (Mexico)",
+		TimeZoneRuleTable_19, 2
+	},
+	{
+		"Eastern Standard Time", 300, TRUE, "(UTC-05:00) Eastern Time (US & Canada)",
+		"Eastern Standard Time", "Eastern Daylight Time",
+		TimeZoneRuleTable_20, 2
+	},
+	{
+		"Haiti Standard Time", 300, TRUE, "(UTC-05:00) Haiti",
+		"Haiti Standard Time", "Haiti Daylight Time",
+		TimeZoneRuleTable_21, 6
+	},
+	{
+		"Cuba Standard Time", 300, TRUE, "(UTC-05:00) Havana",
+		"Cuba Standard Time", "Cuba Daylight Time",
+		TimeZoneRuleTable_22, 11
+	},
+	{
+		"US Eastern Standard Time", 300, TRUE, "(UTC-05:00) Indiana (East)",
+		"US Eastern Standard Time", "US Eastern Daylight Time",
+		TimeZoneRuleTable_23, 2
+	},
+	{
+		"Turks And Caicos Standard Time", 300, TRUE, "(UTC-05:00) Turks and Caicos",
+		"Turks and Caicos Standard Time", "Turks and Caicos Daylight Time",
+		TimeZoneRuleTable_24, 6
+	},
+	{
+		"Paraguay Standard Time", 240, TRUE, "(UTC-04:00) Asuncion",
+		"Paraguay Standard Time", "Paraguay Daylight Time",
+		TimeZoneRuleTable_25, 14
+	},
+	{
+		"Atlantic Standard Time", 240, TRUE, "(UTC-04:00) Atlantic Time (Canada)",
+		"Atlantic Standard Time", "Atlantic Daylight Time",
+		TimeZoneRuleTable_26, 2
+	},
+	{
+		"Venezuela Standard Time", 240, TRUE, "(UTC-04:00) Caracas",
+		"Venezuela Standard Time", "Venezuela Daylight Time",
+		TimeZoneRuleTable_27, 10
+	},
+	{
+		"Central Brazilian Standard Time", 240, TRUE, "(UTC-04:00) Cuiaba",
+		"Central Brazilian Standard Time", "Central Brazilian Daylight Time",
+		TimeZoneRuleTable_28, 37
+	},
+	{
+		"SA Western Standard Time", 240, FALSE, "(UTC-04:00) Georgetown, La Paz, Manaus, San Juan",
+		"SA Western Standard Time", "SA Western Daylight Time",
+		NULL, 0
+	},
+	{
+		"Pacific SA Standard Time", 240, TRUE, "(UTC-04:00) Santiago",
+		"Pacific SA Standard Time", "Pacific SA Daylight Time",
+		TimeZoneRuleTable_30, 10
+	},
+	{
+		"Newfoundland Standard Time", 210, TRUE, "(UTC-03:30) Newfoundland",
+		"Newfoundland Standard Time", "Newfoundland Daylight Time",
+		TimeZoneRuleTable_31, 7
+	},
+	{
+		"Tocantins Standard Time", 180, TRUE, "(UTC-03:00) Araguaina",
+		"Tocantins Standard Time", "Tocantins Daylight Time",
+		TimeZoneRuleTable_32, 2
+	},
+	{
+		"E. South America Standard Time", 180, TRUE, "(UTC-03:00) Brasilia",
+		"E. South America Standard Time", "E. South America Daylight Time",
+		TimeZoneRuleTable_33, 37
+	},
+	{
+		"SA Eastern Standard Time", 180, FALSE, "(UTC-03:00) Cayenne, Fortaleza",
+		"SA Eastern Standard Time", "SA Eastern Daylight Time",
+		NULL, 0
+	},
+	{
+		"Argentina Standard Time", 180, TRUE, "(UTC-03:00) City of Buenos Aires",
+		"Argentina Standard Time", "Argentina Daylight Time",
+		TimeZoneRuleTable_35, 3
+	},
+	{
+		"Greenland Standard Time", 180, TRUE, "(UTC-03:00) Greenland",
+		"Greenland Standard Time", "Greenland Daylight Time",
+		TimeZoneRuleTable_36, 18
+	},
+	{
+		"Montevideo Standard Time", 180, TRUE, "(UTC-03:00) Montevideo",
+		"Montevideo Standard Time", "Montevideo Daylight Time",
+		TimeZoneRuleTable_37, 2
+	},
+	{
+		"Magallanes Standard Time", 180, TRUE, "(UTC-03:00) Punta Arenas",
+		"Magallanes Standard Time", "Magallanes Daylight Time",
+		TimeZoneRuleTable_38, 9
+	},
+	{
+		"Saint Pierre Standard Time", 180, TRUE, "(UTC-03:00) Saint Pierre and Miquelon",
+		"Saint Pierre Standard Time", "Saint Pierre Daylight Time",
+		TimeZoneRuleTable_39, 2
+	},
+	{
+		"Bahia Standard Time", 180, TRUE, "(UTC-03:00) Salvador",
+		"Bahia Standard Time", "Bahia Daylight Time",
+		TimeZoneRuleTable_40, 2
+	},
+	{
+		"UTC-02", 120, FALSE, "(UTC-02:00) Coordinated Universal Time-02",
+		"UTC-02", "UTC-02",
+		NULL, 0
+	},
+	{
+		"Mid-Atlantic Standard Time", 120, TRUE, "(UTC-02:00) Mid-Atlantic - Old",
+		"Mid-Atlantic Standard Time", "Mid-Atlantic Daylight Time",
+		TimeZoneRuleTable_42, 1
+	},
+	{
+		"Azores Standard Time", 60, TRUE, "(UTC-01:00) Azores",
+		"Azores Standard Time", "Azores Daylight Time",
+		TimeZoneRuleTable_43, 1
+	},
+	{
+		"Cape Verde Standard Time", 60, FALSE, "(UTC-01:00) Cabo Verde Is.",
+		"Cabo Verde Standard Time", "Cabo Verde Daylight Time",
+		NULL, 0
+	},
+	{
+		"UTC", 0, FALSE, "(UTC) Coordinated Universal Time",
+		"Coordinated Universal Time", "Coordinated Universal Time",
+		NULL, 0
+	},
+	{
+		"Morocco Standard Time", 0, TRUE, "(UTC+00:00) Casablanca",
+		"Morocco Standard Time", "Morocco Daylight Time",
+		TimeZoneRuleTable_46, 12
+	},
+	{
+		"GMT Standard Time", 0, TRUE, "(UTC+00:00) Dublin, Edinburgh, Lisbon, London",
+		"GMT Standard Time", "GMT Daylight Time",
+		TimeZoneRuleTable_47, 1
+	},
+	{
+		"Greenwich Standard Time", 0, FALSE, "(UTC+00:00) Monrovia, Reykjavik",
+		"Greenwich Standard Time", "Greenwich Daylight Time",
+		NULL, 0
+	},
+	{
+		"W. Europe Standard Time", -60, TRUE, "(UTC+01:00) Amsterdam, Berlin, Bern, Rome, Stockholm, Vienna",
+		"W. Europe Standard Time", "W. Europe Daylight Time",
+		TimeZoneRuleTable_49, 1
+	},
+	{
+		"Central Europe Standard Time", -60, TRUE, "(UTC+01:00) Belgrade, Bratislava, Budapest, Ljubljana, Prague",
+		"Central Europe Standard Time", "Central Europe Daylight Time",
+		TimeZoneRuleTable_50, 1
+	},
+	{
+		"Romance Standard Time", -60, TRUE, "(UTC+01:00) Brussels, Copenhagen, Madrid, Paris",
+		"Romance Standard Time", "Romance Daylight Time",
+		TimeZoneRuleTable_51, 1
+	},
+	{
+		"Sao Tome Standard Time", -60, TRUE, "(UTC+01:00) Sao Tome",
+		"Sao Tome Standard Time", "Sao Tome Daylight Time",
+		TimeZoneRuleTable_52, 2
+	},
+	{
+		"Central European Standard Time", -60, TRUE, "(UTC+01:00) Sarajevo, Skopje, Warsaw, Zagreb",
+		"Central European Standard Time", "Central European Daylight Time",
+		TimeZoneRuleTable_53, 1
+	},
+	{
+		"W. Central Africa Standard Time", -60, FALSE, "(UTC+01:00) West Central Africa",
+		"W. Central Africa Standard Time", "W. Central Africa Daylight Time",
+		NULL, 0
+	},
+	{
+		"Jordan Standard Time", -120, TRUE, "(UTC+02:00) Amman",
+		"Jordan Standard Time", "Jordan Daylight Time",
+		TimeZoneRuleTable_55, 4
+	},
+	{
+		"GTB Standard Time", -120, TRUE, "(UTC+02:00) Athens, Bucharest",
+		"GTB Standard Time", "GTB Daylight Time",
+		TimeZoneRuleTable_56, 1
+	},
+	{
+		"Middle East Standard Time", -120, TRUE, "(UTC+02:00) Beirut",
+		"Middle East Standard Time", "Middle East Daylight Time",
+		TimeZoneRuleTable_57, 13
+	},
+	{
+		"Egypt Standard Time", -120, TRUE, "(UTC+02:00) Cairo",
+		"Egypt Standard Time", "Egypt Daylight Time",
+		TimeZoneRuleTable_58, 7
+	},
+	{
+		"E. Europe Standard Time", -120, TRUE, "(UTC+02:00) Chisinau",
+		"E. Europe Standard Time", "E. Europe Daylight Time",
+		TimeZoneRuleTable_59, 1
+	},
+	{
+		"Syria Standard Time", -120, TRUE, "(UTC+02:00) Damascus",
+		"Syria Standard Time", "Syria Daylight Time",
+		TimeZoneRuleTable_60, 17
+	},
+	{
+		"West Bank Standard Time", -120, TRUE, "(UTC+02:00) Gaza, Hebron",
+		"West Bank Gaza Standard Time", "West Bank Gaza Daylight Time",
+		TimeZoneRuleTable_61, 7
+	},
+	{
+		"South Africa Standard Time", -120, FALSE, "(UTC+02:00) Harare, Pretoria",
+		"South Africa Standard Time", "South Africa Daylight Time",
+		NULL, 0
+	},
+	{
+		"FLE Standard Time", -120, TRUE, "(UTC+02:00) Helsinki, Kyiv, Riga, Sofia, Tallinn, Vilnius",
+		"FLE Standard Time", "FLE Daylight Time",
+		TimeZoneRuleTable_63, 1
+	},
+	{
+		"Israel Standard Time", -120, TRUE, "(UTC+02:00) Jerusalem",
+		"Jerusalem Standard Time", "Jerusalem Daylight Time",
+		TimeZoneRuleTable_64, 20
+	},
+	{
+		"Kaliningrad Standard Time", -120, TRUE, "(UTC+02:00) Kaliningrad",
+		"Russia TZ 1 Standard Time", "Russia TZ 1 Daylight Time",
+		TimeZoneRuleTable_65, 5
+	},
+	{
+		"Sudan Standard Time", -120, TRUE, "(UTC+02:00) Khartoum",
+		"Sudan Standard Time", "Sudan Daylight Time",
+		TimeZoneRuleTable_66, 2
+	},
+	{
+		"Libya Standard Time", -120, TRUE, "(UTC+02:00) Tripoli",
+		"Libya Standard Time", "Libya Daylight Time",
+		TimeZoneRuleTable_67, 2
+	},
+	{
+		"Namibia Standard Time", -120, TRUE, "(UTC+02:00) Windhoek",
+		"Namibia Standard Time", "Namibia Daylight Time",
+		TimeZoneRuleTable_68, 1
+	},
+	{
+		"Arabic Standard Time", -180, TRUE, "(UTC+03:00) Baghdad",
+		"Arabic Standard Time", "Arabic Daylight Time",
+		TimeZoneRuleTable_69, 4
+	},
+	{
+		"Turkey Standard Time", -180, TRUE, "(UTC+03:00) Istanbul",
+		"Turkey Standard Time", "Turkey Daylight Time",
+		TimeZoneRuleTable_70, 7
+	},
+	{
+		"Arab Standard Time", -180, FALSE, "(UTC+03:00) Kuwait, Riyadh",
+		"Arab Standard Time", "Arab Daylight Time",
+		NULL, 0
+	},
+	{
+		"Belarus Standard Time", -180, TRUE, "(UTC+03:00) Minsk",
+		"Belarus Standard Time", "Belarus Daylight Time",
+		TimeZoneRuleTable_72, 2
+	},
+	{
+		"Russian Standard Time", -180, TRUE, "(UTC+03:00) Moscow, St. Petersburg, Volgograd",
+		"Russia TZ 2 Standard Time", "Russia TZ 2 Daylight Time",
+		TimeZoneRuleTable_73, 5
+	},
+	{
+		"E. Africa Standard Time", -180, FALSE, "(UTC+03:00) Nairobi",
+		"E. Africa Standard Time", "E. Africa Daylight Time",
+		NULL, 0
+	},
+	{
+		"Iran Standard Time", -210, TRUE, "(UTC+03:30) Tehran",
+		"Iran Standard Time", "Iran Daylight Time",
+		TimeZoneRuleTable_75, 19
+	},
+	{
+		"Arabian Standard Time", -240, FALSE, "(UTC+04:00) Abu Dhabi, Muscat",
+		"Arabian Standard Time", "Arabian Daylight Time",
+		NULL, 0
+	},
+	{
+		"Astrakhan Standard Time", -240, TRUE, "(UTC+04:00) Astrakhan, Ulyanovsk",
+		"Astrakhan Standard Time", "Astrakhan Daylight Time",
+		TimeZoneRuleTable_77, 5
+	},
+	{
+		"Azerbaijan Standard Time", -240, TRUE, "(UTC+04:00) Baku",
+		"Azerbaijan Standard Time", "Azerbaijan Daylight Time",
+		TimeZoneRuleTable_78, 1
+	},
+	{
+		"Russia Time Zone 3", -240, TRUE, "(UTC+04:00) Izhevsk, Samara",
+		"Russia TZ 3 Standard Time", "Russia TZ 3 Daylight Time",
+		TimeZoneRuleTable_79, 3
+	},
+	{
+		"Mauritius Standard Time", -240, TRUE, "(UTC+04:00) Port Louis",
+		"Mauritius Standard Time", "Mauritius Daylight Time",
+		TimeZoneRuleTable_80, 2
+	},
+	{
+		"Saratov Standard Time", -240, TRUE, "(UTC+04:00) Saratov",
+		"Saratov Standard Time", "Saratov Daylight Time",
+		TimeZoneRuleTable_81, 5
+	},
+	{
+		"Georgian Standard Time", -240, FALSE, "(UTC+04:00) Tbilisi",
+		"Georgian Standard Time", "Georgian Daylight Time",
+		NULL, 0
+	},
+	{
+		"Caucasus Standard Time", -240, TRUE, "(UTC+04:00) Yerevan",
+		"Caucasus Standard Time", "Caucasus Daylight Time",
+		TimeZoneRuleTable_83, 1
+	},
+	{
+		"Afghanistan Standard Time", -270, FALSE, "(UTC+04:30) Kabul",
+		"Afghanistan Standard Time", "Afghanistan Daylight Time",
+		NULL, 0
+	},
+	{
+		"West Asia Standard Time", -300, FALSE, "(UTC+05:00) Ashgabat, Tashkent",
+		"West Asia Standard Time", "West Asia Daylight Time",
+		NULL, 0
+	},
+	{
+		"Ekaterinburg Standard Time", -300, TRUE, "(UTC+05:00) Ekaterinburg",
+		"Russia TZ 4 Standard Time", "Russia TZ 4 Daylight Time",
+		TimeZoneRuleTable_86, 5
+	},
+	{
+		"Pakistan Standard Time", -300, TRUE, "(UTC+05:00) Islamabad, Karachi",
+		"Pakistan Standard Time", "Pakistan Daylight Time",
+		TimeZoneRuleTable_87, 2
+	},
+	{
+		"India Standard Time", -330, FALSE, "(UTC+05:30) Chennai, Kolkata, Mumbai, New Delhi",
+		"India Standard Time", "India Daylight Time",
+		NULL, 0
+	},
+	{
+		"Sri Lanka Standard Time", -330, FALSE, "(UTC+05:30) Sri Jayawardenepura",
+		"Sri Lanka Standard Time", "Sri Lanka Daylight Time",
+		NULL, 0
+	},
+	{
+		"Nepal Standard Time", -345, FALSE, "(UTC+05:45) Kathmandu",
+		"Nepal Standard Time", "Nepal Daylight Time",
+		NULL, 0
+	},
+	{
+		"Central Asia Standard Time", -360, FALSE, "(UTC+06:00) Astana",
+		"Central Asia Standard Time", "Central Asia Daylight Time",
+		NULL, 0
+	},
+	{
+		"Bangladesh Standard Time", -360, TRUE, "(UTC+06:00) Dhaka",
+		"Bangladesh Standard Time", "Bangladesh Daylight Time",
+		TimeZoneRuleTable_92, 1
+	},
+	{
+		"Omsk Standard Time", -360, TRUE, "(UTC+06:00) Omsk",
+		"Omsk Standard Time", "Omsk Daylight Time",
+		TimeZoneRuleTable_93, 5
+	},
+	{
+		"Myanmar Standard Time", -390, FALSE, "(UTC+06:30) Yangon (Rangoon)",
+		"Myanmar Standard Time", "Myanmar Daylight Time",
+		NULL, 0
+	},
+	{
+		"SE Asia Standard Time", -420, FALSE, "(UTC+07:00) Bangkok, Hanoi, Jakarta",
+		"SE Asia Standard Time", "SE Asia Daylight Time",
+		NULL, 0
+	},
+	{
+		"Altai Standard Time", -420, TRUE, "(UTC+07:00) Barnaul, Gorno-Altaysk",
+		"Altai Standard Time", "Altai Daylight Time",
+		TimeZoneRuleTable_96, 5
+	},
+	{
+		"W. Mongolia Standard Time", -420, TRUE, "(UTC+07:00) Hovd",
+		"W. Mongolia Standard Time", "W. Mongolia Daylight Time",
+		TimeZoneRuleTable_97, 3
+	},
+	{
+		"North Asia Standard Time", -420, TRUE, "(UTC+07:00) Krasnoyarsk",
+		"Russia TZ 6 Standard Time", "Russia TZ 6 Daylight Time",
+		TimeZoneRuleTable_98, 5
+	},
+	{
+		"N. Central Asia Standard Time", -420, TRUE, "(UTC+07:00) Novosibirsk",
+		"Novosibirsk Standard Time", "Novosibirsk Daylight Time",
+		TimeZoneRuleTable_99, 5
+	},
+	{
+		"Tomsk Standard Time", -420, TRUE, "(UTC+07:00) Tomsk",
+		"Tomsk Standard Time", "Tomsk Daylight Time",
+		TimeZoneRuleTable_100, 5
+	},
+	{
+		"China Standard Time", -480, FALSE, "(UTC+08:00) Beijing, Chongqing, Hong Kong, Urumqi",
+		"China Standard Time", "China Daylight Time",
+		NULL, 0
+	},
+	{
+		"North Asia East Standard Time", -480, TRUE, "(UTC+08:00) Irkutsk",
+		"Russia TZ 7 Standard Time", "Russia TZ 7 Daylight Time",
+		TimeZoneRuleTable_102, 5
+	},
+	{
+		"Singapore Standard Time", -480, FALSE, "(UTC+08:00) Kuala Lumpur, Singapore",
+		"Malay Peninsula Standard Time", "Malay Peninsula Daylight Time",
+		NULL, 0
+	},
+	{
+		"W. Australia Standard Time", -480, TRUE, "(UTC+08:00) Perth",
+		"W. Australia Standard Time", "W. Australia Daylight Time",
+		TimeZoneRuleTable_104, 4
+	},
+	{
+		"Taipei Standard Time", -480, FALSE, "(UTC+08:00) Taipei",
+		"Taipei Standard Time", "Taipei Daylight Time",
+		NULL, 0
+	},
+	{
+		"Ulaanbaatar Standard Time", -480, TRUE, "(UTC+08:00) Ulaanbaatar",
+		"Ulaanbaatar Standard Time", "Ulaanbaatar Daylight Time",
+		TimeZoneRuleTable_106, 2
+	},
+	{
+		"Aus Central W. Standard Time", -525, FALSE, "(UTC+08:45) Eucla",
+		"Aus Central W. Standard Time", "Aus Central W. Daylight Time",
+		NULL, 0
+	},
+	{
+		"Transbaikal Standard Time", -540, TRUE, "(UTC+09:00) Chita",
+		"Transbaikal Standard Time", "Transbaikal Daylight Time",
+		TimeZoneRuleTable_108, 7
+	},
+	{
+		"Tokyo Standard Time", -540, FALSE, "(UTC+09:00) Osaka, Sapporo, Tokyo",
+		"Tokyo Standard Time", "Tokyo Daylight Time",
+		NULL, 0
+	},
+	{
+		"North Korea Standard Time", -540, TRUE, "(UTC+09:00) Pyongyang",
+		"North Korea Standard Time", "North Korea Daylight Time",
+		TimeZoneRuleTable_110, 4
+	},
+	{
+		"Korea Standard Time", -540, FALSE, "(UTC+09:00) Seoul",
+		"Korea Standard Time", "Korea Daylight Time",
+		NULL, 0
+	},
+	{
+		"Yakutsk Standard Time", -540, TRUE, "(UTC+09:00) Yakutsk",
+		"Russia TZ 8 Standard Time", "Russia TZ 8 Daylight Time",
+		TimeZoneRuleTable_112, 5
+	},
+	{
+		"Cen. Australia Standard Time", -570, TRUE, "(UTC+09:30) Adelaide",
+		"Cen. Australia Standard Time", "Cen. Australia Daylight Time",
+		TimeZoneRuleTable_113, 2
+	},
+	{
+		"AUS Central Standard Time", -570, FALSE, "(UTC+09:30) Darwin",
+		"AUS Central Standard Time", "AUS Central Daylight Time",
+		NULL, 0
+	},
+	{
+		"E. Australia Standard Time", -600, FALSE, "(UTC+10:00) Brisbane",
+		"E. Australia Standard Time", "E. Australia Daylight Time",
+		NULL, 0
+	},
+	{
+		"AUS Eastern Standard Time", -600, TRUE, "(UTC+10:00) Canberra, Melbourne, Sydney",
+		"AUS Eastern Standard Time", "AUS Eastern Daylight Time",
+		TimeZoneRuleTable_116, 2
+	},
+	{
+		"West Pacific Standard Time", -600, FALSE, "(UTC+10:00) Guam, Port Moresby",
+		"West Pacific Standard Time", "West Pacific Daylight Time",
+		NULL, 0
+	},
+	{
+		"Tasmania Standard Time", -600, TRUE, "(UTC+10:00) Hobart",
+		"Tasmania Standard Time", "Tasmania Daylight Time",
+		TimeZoneRuleTable_118, 2
+	},
+	{
+		"Vladivostok Standard Time", -600, TRUE, "(UTC+10:00) Vladivostok",
+		"Russia TZ 9 Standard Time", "Russia TZ 9 Daylight Time",
+		TimeZoneRuleTable_119, 5
+	},
+	{
+		"Lord Howe Standard Time", -630, TRUE, "(UTC+10:30) Lord Howe Island",
+		"Lord Howe Standard Time", "Lord Howe Daylight Time",
+		TimeZoneRuleTable_120, 4
+	},
+	{
+		"Bougainville Standard Time", -660, TRUE, "(UTC+11:00) Bougainville Island",
+		"Bougainville Standard Time", "Bougainville Daylight Time",
+		TimeZoneRuleTable_121, 2
+	},
+	{
+		"Russia Time Zone 10", -660, TRUE, "(UTC+11:00) Chokurdakh",
+		"Russia TZ 10 Standard Time", "Russia TZ 10 Daylight Time",
+		TimeZoneRuleTable_122, 5
+	},
+	{
+		"Magadan Standard Time", -660, TRUE, "(UTC+11:00) Magadan",
+		"Magadan Standard Time", "Magadan Daylight Time",
+		TimeZoneRuleTable_123, 7
+	},
+	{
+		"Norfolk Standard Time", -660, TRUE, "(UTC+11:00) Norfolk Island",
+		"Norfolk Standard Time", "Norfolk Daylight Time",
+		TimeZoneRuleTable_124, 2
+	},
+	{
+		"Sakhalin Standard Time", -660, TRUE, "(UTC+11:00) Sakhalin",
+		"Sakhalin Standard Time", "Sakhalin Daylight Time",
+		TimeZoneRuleTable_125, 5
+	},
+	{
+		"Central Pacific Standard Time", -660, FALSE, "(UTC+11:00) Solomon Is., New Caledonia",
+		"Central Pacific Standard Time", "Central Pacific Daylight Time",
+		NULL, 0
+	},
+	{
+		"Russia Time Zone 11", -720, TRUE, "(UTC+12:00) Anadyr, Petropavlovsk-Kamchatsky",
+		"Russia TZ 11 Standard Time", "Russia TZ 11 Daylight Time",
+		TimeZoneRuleTable_127, 3
+	},
+	{
+		"New Zealand Standard Time", -720, TRUE, "(UTC+12:00) Auckland, Wellington",
+		"New Zealand Standard Time", "New Zealand Daylight Time",
+		TimeZoneRuleTable_128, 3
+	},
+	{
+		"UTC+12", -720, FALSE, "(UTC+12:00) Coordinated Universal Time+12",
+		"UTC+12", "UTC+12",
+		NULL, 0
+	},
+	{
+		"Fiji Standard Time", -720, TRUE, "(UTC+12:00) Fiji",
+		"Fiji Standard Time", "Fiji Daylight Time",
+		TimeZoneRuleTable_130, 11
+	},
+	{
+		"Kamchatka Standard Time", -720, TRUE, "(UTC+12:00) Petropavlovsk-Kamchatsky - Old",
+		"Kamchatka Standard Time", "Kamchatka Daylight Time",
+		TimeZoneRuleTable_131, 1
+	},
+	{
+		"Chatham Islands Standard Time", -765, TRUE, "(UTC+12:45) Chatham Islands",
+		"Chatham Islands Standard Time", "Chatham Islands Daylight Time",
+		TimeZoneRuleTable_132, 3
+	},
+	{
+		"UTC+13", -780, FALSE, "(UTC+13:00) Coordinated Universal Time+13",
+		"UTC+13", "UTC+13",
+		NULL, 0
+	},
+	{
+		"Tonga Standard Time", -780, TRUE, "(UTC+13:00) Nuku'alofa",
+		"Tonga Standard Time", "Tonga Daylight Time",
+		TimeZoneRuleTable_134, 2
+	},
+	{
+		"Samoa Standard Time", -780, TRUE, "(UTC+13:00) Samoa",
+		"Samoa Standard Time", "Samoa Daylight Time",
+		TimeZoneRuleTable_135, 4
+	},
+	{
+		"Line Islands Standard Time", -840, FALSE, "(UTC+14:00) Kiritimati Island",
+		"Line Islands Standard Time", "Line Islands Daylight Time",
+		NULL, 0
+	}
+};
+
